Made user profile page a boomerang
'Boomerang page' is a term I made up that means a page that logs its path in $_SESSION['fromPage'] before redirecting the user here, e.g. the user profile page, if accessed by an unsigned-in user, redirects them here, prompting them to sign in, and upon successful login, the user will be redirected back to the user profile page, like a boomerang!